rename: file naming format to be consistent
